+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v1.19.0 (2023-02-XX)
 - Allow user to subscribe to a recruitment process
+- Allow user to subscribe to an offer
 
 # v1.18.0 (2023-01-31)
 - Fix interview page sort order